+++ typings/handlers/ButtonsHandler.d.ts
@@ -3,6 +3,6 @@ export declare class ButtonsHandler {
     private client;
     private dir;
     constructor(client: ShewenyClient, dir: string);
-    registerAll(): Promise<import("collection-data").Collection<string, import("../typescript/interfaces/Event").Event>>;
+    registerAll(): Promise<import("collection-data").Collection<string, import("../typescript/interfaces/interfaces").Event>>;
     readDirAndPush(d: string): Promise<Array<string>>;
 }
