+++ typings/structures/Button.d.ts
@@ -1,6 +1,12 @@
-import { IButtonMeta } from "../typescript/interfaces/Button";
 import type { EmojiIdentifierResolvable } from "discord.js";
 import type { ShewenyClient } from "../index";
+export interface IButtonMeta {
+    description?: string;
+    style: "PRIMARY" | "SECONDARY" | "SUCCESS" | "DANGER" | "LINK";
+    disabled?: boolean;
+    emoji?: EmojiIdentifierResolvable;
+    label?: string;
+}
 export declare class Button {
     client: any;
     path?: string;
