+++ typings/structures/Event.d.ts
@@ -1,5 +1,9 @@
-import { IEventMeta } from "../typescript/interfaces/Event";
 import type { ShewenyClient } from "../index";
+interface IEventMeta {
+    name: string;
+    description: string;
+    once: boolean;
+}
 /**
  * Represent a event
  * @class
@@ -32,3 +36,4 @@ export declare class Event {
      */
     register(): Promise<any>;
 }
+export {};
