--- typings/typescript/interfaces/CommandHandler.d.ts
@@ -1,5 +1,0 @@
-export interface ICommandHandlerOptions {
-    type?: "SLASH_COMMANDS" | "MESSAGE_COMMANDS";
-    directory: string;
-    prefix?: string;
-}
